+++ src/main/java/de/samply/db/model/ProjectBridgeheadUser.java
@@ -1,11 +1,14 @@
 package de.samply.db.model;
 
 
+import de.samply.project.state.UserProjectState;
 import de.samply.user.roles.ProjectRole;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.time.Instant;
 
 @Entity
 @Table(name = "project_bridgehead_user", schema = "samply")
@@ -29,5 +32,12 @@ public class ProjectBridgeheadUser {
     @Column(name = "project_role", nullable = false)
     @Enumerated(EnumType.STRING)
     private ProjectRole projectRole;
+
+    @Column(name = "project_state", nullable = false)
+    @Enumerated(EnumType.STRING)
+    private UserProjectState projectState = UserProjectState.CREATED;
+
+    @Column(name = "modified_at", nullable = false)
+    private Instant modifiedAt = Instant.now();
 
 }