+++ src/main/java/de/samply/db/model/ProjectDocument.java
@@ -34,7 +34,7 @@ public class ProjectDocument {
     private String url;
 
     @Column(name = "created_at", nullable = false)
-    private Instant createdAt;
+    private Instant createdAt = Instant.now();
 
     @Column(name = "bridgehead")
     private String bridgehead;