+++ src/main/java/de/samply/user/UserService.java
@@ -2,10 +2,13 @@ package de.samply.user;
 
 import de.samply.db.model.*;
 import de.samply.db.repository.*;
+import de.samply.project.state.UserProjectState;
+import de.samply.security.SessionUser;
 import de.samply.user.roles.ProjectRole;
 import jakarta.validation.constraints.NotNull;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
 import java.util.Optional;
 
 @Service
@@ -16,17 +19,20 @@ public class UserService {
     private final ProjectBridgeheadUserRepository projectBridgeheadUserRepository;
     private final ProjectRepository projectRepository;
     private final ProjectBridgeheadRepository projectBridgeheadRepository;
+    private final SessionUser sessionUser;
 
     public UserService(BridgeheadAdminUserRepository bridgeheadAdminUserRepository,
                        ProjectManagerAdminUserRepository projectManagerAdminUserRepository,
                        ProjectBridgeheadUserRepository projectBridgeheadUserRepository,
                        ProjectRepository projectRepository,
-                       ProjectBridgeheadRepository projectBridgeheadRepository) {
+                       ProjectBridgeheadRepository projectBridgeheadRepository,
+                       SessionUser sessionUser) {
         this.bridgeheadAdminUserRepository = bridgeheadAdminUserRepository;
         this.projectManagerAdminUserRepository = projectManagerAdminUserRepository;
         this.projectBridgeheadUserRepository = projectBridgeheadUserRepository;
         this.projectRepository = projectRepository;
         this.projectBridgeheadRepository = projectBridgeheadRepository;
+        this.sessionUser = sessionUser;
     }
 
     public BridgeheadAdminUser createBridgeheadAdminUserIfNotExists(@NotNull String email, @NotNull String bridgehead) {
@@ -86,8 +92,40 @@ public class UserService {
             projectBridgeheadUser.setEmail(email);
             projectBridgeheadUser.setProjectRole(projectRole);
             projectBridgeheadUser.setProjectBridgehead(projectBridgehead.get());
+            projectBridgeheadUser.setProjectState(UserProjectState.CREATED);
+            projectBridgeheadUser.setModifiedAt(Instant.now());
             this.projectBridgeheadUserRepository.save(projectBridgeheadUser);
         }
+    }
+
+    public void acceptProject(@NotNull String projectCode, @NotNull String bridgehead) throws UserServiceException {
+        changeProjectState(projectCode, bridgehead, UserProjectState.ACCEPTED);
+    }
+
+    public void rejectProject(@NotNull String projectCode, @NotNull String bridgehead) throws UserServiceException {
+        changeProjectState(projectCode, bridgehead, UserProjectState.REJECTED);
+    }
+
+    public void requestChangesInProject(@NotNull String projectCode, @NotNull String bridgehead) throws UserServiceException {
+        changeProjectState(projectCode, bridgehead, UserProjectState.REQUEST_CHANGES);
+    }
+
+    private void changeProjectState(@NotNull String projectCode, @NotNull String bridgehead, @NotNull UserProjectState state) throws UserServiceException {
+        Optional<Project> project = projectRepository.findByCode(projectCode);
+        if (project.isEmpty()) {
+            throw new UserServiceException("Project " + projectCode + " not found");
+        }
+        Optional<ProjectBridgehead> projectBridgehead = projectBridgeheadRepository.findFirstByBridgeheadAndProject(bridgehead, project.get());
+        if (projectBridgehead.isEmpty()) {
+            throw new UserServiceException("Project " + projectCode + " for bridgehead " + bridgehead + " not found");
+        }
+        Optional<ProjectBridgeheadUser> projectBridgeheadUser = projectBridgeheadUserRepository.getFirstByEmailAndProjectBridgeheadOrderByModifiedAtDesc(sessionUser.getEmail(), projectBridgehead.get());
+        if (projectBridgeheadUser.isEmpty()) {
+            throw new UserServiceException("Project " + projectCode + " for bridgehead " + bridgehead + " and user " + sessionUser.getEmail());
+        }
+        projectBridgeheadUser.get().setProjectState(state);
+        projectBridgeheadUser.get().setModifiedAt(Instant.now());
+        projectBridgeheadUserRepository.save(projectBridgeheadUser.get());
     }
 
 }